fix: Small Fixes Made To Bank Library. Additions To Param.
This update fixes some minor issues found in the Bank library. Notably, it changes the return type of IRVBankEntry properties like Offset, OriginalSize, TimeStamp and others from long to uint. Furthermore, IRVBankEntry's method CalculateLength now returns uint as well.

In addition to that, various changes were made to support new features in the Param library, such as the ability to obtain array values. Some additions were also made to provide more functionality such as creating, adding, and removing variables.

The update also includes a new feature for reading param files located in ReadParamFile method inside Src/BisUtils.Param/Model/ParamFile.cs.

Finally, the importance of properties like OriginalSize, Offset, TimeStamp in RVBankDirectory.cs has been changed to uint, for consistency with other parts of the system. The commit also includes minor code cleanups and improvements across multiple files.

+++ src/BisUtils.RVBank/Model/Entry/RVBankDataEntry.cs
@@ -55,10 +55,10 @@ public class RVBankDataEntry : RVBankEntry, IRVBankDataEntry
         IRVBankDirectory parent,
         string fileName,
         RVBankEntryMime mime,
-        int originalSize,
-        int offset,
-        int timeStamp,
-        int dataSize
+        uint originalSize,
+        uint offset,
+        uint timeStamp,
+        uint dataSize
     ) : base(file, parent, fileName, mime, originalSize, offset, timeStamp, dataSize)
     {
     }
@@ -69,11 +69,11 @@ public class RVBankDataEntry : RVBankEntry, IRVBankDataEntry
         IRVBankDirectory parent,
         string fileName,
         RVBankEntryMime mime,
-        long offset,
-        long timeStamp,
+        uint offset,
+        uint timeStamp,
         Stream entryData,
         RVBankDataType packingMethod
-    ) : base(file, parent, fileName, mime, entryData.Length, offset, timeStamp, 0) =>
+    ) : base(file, parent, fileName, mime, (uint) entryData.Length, offset, timeStamp, 0) =>
         PackingMethod = packingMethod;
 
     protected sealed override void OnChangesMade(object? sender, EventArgs? e) => base.OnChangesMade(sender, e);
@@ -87,7 +87,7 @@ public class RVBankDataEntry : RVBankEntry, IRVBankDataEntry
         }
     }
 
-    public void SynchronizeMetaWithStream() => OriginalSize = (int)EntryData.Length;
+    public void SynchronizeMetaWithStream() => OriginalSize = (uint)EntryData.Length;
 
     public void ExpandDirectoryStructure()
     {
@@ -273,10 +273,10 @@ public class RVBankDataEntry : RVBankEntry, IRVBankDataEntry
     {
         LastResult = base.Debinarize(reader, options);
         EntryMime = (RVBankEntryMime)reader.ReadInt32(); // TODO WARN/ERROR then recover
-        OriginalSize = reader.ReadInt32();
-        TimeStamp = reader.ReadInt32();
-        Offset = reader.ReadInt32();
-        DataSize = reader.ReadInt32();
+        OriginalSize = reader.ReadUInt32();
+        TimeStamp = reader.ReadUInt32();
+        Offset = reader.ReadUInt32();
+        DataSize = reader.ReadUInt32();
 
         return LastResult;
     }
